Get Nodesync to import textures and materials to the specified folder
The call to GetDefaultMaterialAndTextureCookMode() always returns CookToTemp,
so both the textures and materials would always be cooked to a temp folder.
However, NodeSync provides a parameter letting you select which folder
the texture/material will be saved to (the Unreal Asset Import Folder
parameter), and cooking to temp means that parameter isn't respected.
So we remove the call to GetDefaultMaterialAndTextureCookMode(),
and instead just let MyPackageParams copy the PackageMode of InPackageParams
(which will be set to either Bake when importing using NodeSync, and
CookToTemp if the texture is attached to an unbaked HDA).