Fix incorrect determination of whether a texture has alpha
The previous for loop assumed the data was packed as RGBA. It checked
every four bytes for a value lower than 1.0, and if there was one,
marked the texture as translucent. However, this would work incorrectly
on RGB, UV, and Mono packing, checking a non-alpha byte for alpha. For
example, the fourth byte of RGB packing would be the R value of the second
pixel: RGB[R]GB. Most likely this byte would not be 1.0, and the texture
would incorrectly be marked as translucent.