Implement COP texture creation method with Unreal implementation
This commit implements a new HAPI method,
HAPI_CreateCOPImage. This method takes in an
array of raw pixel data, and creates a copnet
containing a cop node containing that pixel data.

This commit also implements the ability send
textures from Unreal to Houdini (through the above
HAPI method). Simply right click an Unreal texture
in the Content Browser, and select "Send To Houdini".

For now, since there is no way to save data in a
COP node, we use a workaround. The workaround creates
a heightfield for each channel (R, G, B, A) of the image,
and uses the SOP import COP to bring the volume into
the copnet. It then converts the SOP into a texture
using the geotolayer COP node, and merges the four
channels. In the future, once we have a way to save
data to a COP node, the method should be adjusted to
use that instead of the heightfield workaround.

Right now the method is called HAPI_CreateCOPImage,
but I discussed with Damien about multithreading usage.
In the future we'd like to support splitting the image up
into parts and have each thread make a call to the method
to write one part. So we might be rename the method to
HAPI_SetCOPImageData. The start and length parameters are
supposed to be used for this purpose, but as of now don't
do anything.